Add namespace support for both kernels and tensors.
This includes updates to code generation and the API.

It also includes a NamespacedGenerator which wraps a Generator
and sets a namespace for all kernels generated with it.
The XML/Json parse functions support namespaces as well.

Furthermore, when closing a namespace a comment is included which
makes clear, which namespace is closed.